split import and export ops
