+++ OpenKeychain/src/main/java/org/sufficientlysecure/keychain/service/ImportKeyringParcel.java
@@ -24,7 +24,7 @@ import org.sufficientlysecure.keychain.keyimport.ParcelableKeyRing;
 
 import java.util.ArrayList;
 
-public class ImportKeyringParcel extends ImportExportParcel {
+public class ImportKeyringParcel implements Parcelable {
     // if null, keys are expected to be read from a cache file in ImportExportOperations
     public ArrayList<ParcelableKeyRing> mKeyList;
     public String mKeyserver; // must be set if keys are to be imported from a keyserver
